noop rename exploreRow -> discovery
