+++ src/components/Discovery.vue
